Fix more build warnings
Remove all references to SyntaxGenerator from analyzer class
